+++ spinifex/daemon/daemon.go
@@ -1,7 +1,6 @@
 package daemon
 
 import (
-	"bufio"
 	"context"
 	"crypto/sha256"
 	"crypto/tls"
@@ -54,7 +53,6 @@ import (
 	"github.com/mulgadc/spinifex/spinifex/types"
 	"github.com/mulgadc/spinifex/spinifex/utils"
 	"github.com/mulgadc/spinifex/spinifex/vm"
-	"github.com/mulgadc/viperblock/viperblock"
 	"github.com/nats-io/nats.go"
 )
 
@@ -1289,7 +1287,7 @@ func (d *Daemon) restoreInstances() {
 					return
 				}
 				slog.Info("Launching instance (recovery)", "instance", inst.ID)
-				if err := d.LaunchInstance(inst); err != nil {
+				if err := d.vmMgr.Run(inst); err != nil {
 					slog.Error("Failed to launch instance during recovery", "instanceId", inst.ID, "err", err)
 					d.markInstanceFailed(inst, "recovery_launch_failed")
 				}
@@ -1347,7 +1345,7 @@ func (d *Daemon) areVolumeSocketsValid(instance *vm.VM) bool {
 // after a daemon restart. This bypasses the state machine since recovery is not a
 // normal state transition.
 func (d *Daemon) reconnectInstance(instance *vm.VM) error {
-	if err := d.CreateQMPClient(instance); err != nil {
+	if err := d.vmMgr.AttachQMP(instance); err != nil {
 		return fmt.Errorf("failed to reconnect QMP: %w", err)
 	}
 
@@ -1894,190 +1892,6 @@ func (d *Daemon) setupShutdown() {
 	})
 }
 
-func (d *Daemon) CreateQMPClient(instance *vm.VM) (err error) {
-	// Create a new QMP client to communicate with the instance
-	instance.QMPClient, err = qmp.NewQMPClient(instance.Config.QMPSocket)
-
-	if err != nil {
-		slog.Error("Could not connect to QMP")
-		return err
-	}
-
-	// Send qmp_capabilities handshake to init
-	_, err = d.SendQMPCommand(instance.QMPClient, qmp.QMPCommand{Execute: "qmp_capabilities"}, instance.ID)
-	if err != nil {
-		slog.Error("Failed QMP capabilities handshake", "err", err)
-		return err
-	}
-
-	// Simple heartbeat to confirm QMP and the instance is running / healthy
-	go func() {
-		for {
-			time.Sleep(30 * time.Second)
-
-			// Check if instance is in a terminal or transitional state - exit heartbeat
-			var status vm.InstanceState
-			d.vmMgr.Inspect(instance, func(v *vm.VM) { status = v.Status })
-
-			if status == vm.StateStopping || status == vm.StateStopped || status == vm.StateShuttingDown || status == vm.StateTerminated || status == vm.StateError {
-				slog.Info("QMP heartbeat exiting - instance not running", "instance", instance.ID, "status", status)
-
-				// Close the QMP client connection if it exists
-				if instance.QMPClient != nil && instance.QMPClient.Conn != nil {
-					if err := instance.QMPClient.Conn.Close(); err != nil {
-						slog.Error("Failed to close QMP connection", "instance", instance.ID, "err", err)
-					}
-				}
-				return
-			}
-
-			slog.Debug("QMP heartbeat", "instance", instance.ID)
-			qmpStatus, err := d.SendQMPCommand(instance.QMPClient, qmp.QMPCommand{Execute: "query-status"}, instance.ID)
-
-			if err != nil {
-				slog.Warn("QMP heartbeat failed", "instance", instance.ID, "err", err)
-				// Don't exit on transient errors - let the status check above handle terminal states
-				continue
-			}
-
-			slog.Debug("QMP status", "instance", instance.ID, "status", string(qmpStatus.Return))
-		}
-	}()
-
-	return nil
-}
-
-// launchStillValid returns true while LaunchInstance may continue setting up
-// resources for instance. Returns false if a concurrent terminate has flipped
-// status out of pending/stopped/provisioning — at that point the terminate
-// goroutine owns cleanup and LaunchInstance must bail without further side
-// effects. Logs the abort reason at Info so it's visible without polluting
-// error logs (terminate-during-pending is an expected user-driven race).
-func (d *Daemon) launchStillValid(instance *vm.VM) bool {
-	var status vm.InstanceState
-	d.vmMgr.Inspect(instance, func(v *vm.VM) { status = v.Status })
-	if status == vm.StatePending || status == vm.StateStopped || status == vm.StateProvisioning {
-		return true
-	}
-	slog.Info("Launch aborted by concurrent terminate", "instanceId", instance.ID, "status", string(status))
-	return false
-}
-
-func (d *Daemon) LaunchInstance(instance *vm.VM) (err error) {
-	// Abort if instance is no longer in a launchable state. A concurrent
-	// terminate request that flipped status to shutting-down/terminated
-	// owns the cleanup lifecycle; this path is an expected race outcome,
-	// not an error, so return nil to avoid the spurious
-	// "Failed to transition instance to running" ERROR log.
-	if !d.launchStillValid(instance) {
-		return nil
-	}
-
-	// First, confirm if the instance is already running
-	pid, _ := utils.ReadPidFile(instance.ID)
-
-	if pid > 0 {
-		process, err := os.FindProcess(pid)
-		if err != nil {
-			return err
-		}
-
-		// Send a 0 signal to confirm process is running
-		err = process.Signal(syscall.Signal(0))
-		if err == nil {
-			slog.Error("Instance is already running", "InstanceID", instance.ID, "pid", pid)
-			return errors.New("instance is already running")
-		}
-	}
-
-	// Loop through each volume in volumes
-	err = d.MountVolumes(instance)
-
-	if err != nil {
-		slog.Error("Failed to mount volumes", "err", err)
-		return err
-	}
-
-	// Re-check status — MountVolumes can take 30+s on cold AMIs (NBD
-	// clone), and a terminate can race in during that window. Skip the
-	// remaining setup so the concurrent terminate goroutine doesn't fight
-	// SetupTapDevice and leak resources. PR1b.1's idempotency safety net
-	// catches anything that does slip through.
-	if !d.launchStillValid(instance) {
-		return nil
-	}
-
-	// Step 6: Launch the instance via QEMU/KVM
-	err = d.StartInstance(instance)
-
-	if err != nil {
-		slog.Error("Failed to launch instance", "err", err)
-		return err
-	}
-
-	// Step 7: Create QMP client to communicate with the instance
-	err = d.CreateQMPClient(instance)
-
-	if err != nil {
-		slog.Error("Failed to create QMP client", "err", err)
-		return err
-	}
-
-	// Step 8: Subscribe to start/stop/shutdown events
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	// Unsubscribe any existing subscriptions (e.g. from restoreInstances for stopped instances)
-	if existing, ok := d.natsSubscriptions[instance.ID]; ok {
-		_ = existing.Unsubscribe()
-	}
-	consoleSubKey := instance.ID + ".console"
-	if existing, ok := d.natsSubscriptions[consoleSubKey]; ok {
-		_ = existing.Unsubscribe()
-	}
-
-	d.natsSubscriptions[instance.ID], err = d.natsConn.Subscribe(fmt.Sprintf("ec2.cmd.%s", instance.ID), d.handleEC2Events)
-	if err != nil {
-		slog.Error("failed to subscribe to NATS", "err", err)
-		return err
-	}
-
-	d.natsSubscriptions[consoleSubKey], err = d.natsConn.Subscribe(fmt.Sprintf("ec2.%s.GetConsoleOutput", instance.ID), d.handleEC2GetConsoleOutput)
-	if err != nil {
-		slog.Error("failed to subscribe to console output NATS topic", "err", err)
-		return err
-	}
-
-	// Step 9: Update the instance metadata for running state and volume attached
-	d.vmMgr.Insert(instance)
-
-	// Final race check — QEMU is up, but if a terminate fired during
-	// StartInstance/CreateQMPClient, the concurrent goroutine has already
-	// transitioned status to shutting-down. Attempting StateRunning here
-	// would log a spurious error; let the terminate cleanup own teardown.
-	if !d.launchStillValid(instance) {
-		return nil
-	}
-
-	if err := d.TransitionState(instance, vm.StateRunning); err != nil {
-		slog.Error("Failed to transition instance to running", "instanceId", instance.ID, "err", err)
-		return err
-	}
-
-	// Step 10: Mark boot volumes as "in-use" now that instance is confirmed running
-	instance.EBSRequests.Mu.Lock()
-	for _, ebsReq := range instance.EBSRequests.Requests {
-		if ebsReq.Boot {
-			if err := d.volumeService.UpdateVolumeState(ebsReq.Name, "in-use", instance.ID, ""); err != nil {
-				slog.Error("Failed to update volume state to in-use", "volumeId", ebsReq.Name, "err", err)
-			}
-		}
-	}
-	instance.EBSRequests.Mu.Unlock()
-
-	return nil
-}
-
 // markInstanceFailed updates an instance status to indicate a failure during launch,
 // then completes the termination lifecycle in the background so the instance
 // reaches terminated state and doesn't get stuck in shutting-down.
@@ -2202,377 +2016,6 @@ func (d *Daemon) startPendingWatchdog() {
 			}
 		}
 	}()
-}
-
-func (d *Daemon) StartInstance(instance *vm.VM) error {
-	pidFile, err := utils.GeneratePidFile(instance.ID)
-
-	if err != nil {
-		slog.Error("Failed to generate PID file", "err", err)
-		return err
-	}
-
-	instanceType := d.resourceMgr.instanceTypes[instance.InstanceType]
-	if instanceType == nil {
-		return fmt.Errorf("instance type %s not found", instance.InstanceType)
-	}
-
-	vCPUs := int(instanceTypeVCPUs(instanceType))
-	memoryMiB := instanceTypeMemoryMiB(instanceType)
-	architecture := "x86_64"
-	if instanceType.ProcessorInfo != nil && len(instanceType.ProcessorInfo.SupportedArchitectures) > 0 && instanceType.ProcessorInfo.SupportedArchitectures[0] != nil {
-		architecture = *instanceType.ProcessorInfo.SupportedArchitectures[0]
-	}
-
-	// Console log + serial socket paths (serial output capture + admin access via socat)
-	runtimeDir := utils.RuntimeDir()
-	consoleLogPath := filepath.Join(runtimeDir, fmt.Sprintf("console-%s.log", instance.ID))
-	serialSocket := filepath.Join(runtimeDir, fmt.Sprintf("serial-%s.sock", instance.ID))
-
-	instance.Config = buildBaseVMConfig(instance.ID, pidFile, consoleLogPath, serialSocket, architecture, vCPUs, int(memoryMiB))
-
-	// Build QEMU drives from EBS volume requests.
-	instance.EBSRequests.Mu.Lock()
-	drives, iothreads, devices, err := buildDrives(instance.EBSRequests.Requests, vCPUs)
-	instance.EBSRequests.Mu.Unlock()
-	if err != nil {
-		return err
-	}
-	instance.Config.Drives = append(instance.Config.Drives, drives...)
-	instance.Config.IOThreads = append(instance.Config.IOThreads, iothreads...)
-	instance.Config.Devices = append(instance.Config.Devices, devices...)
-
-	// VPC tap networking vs user-mode fallback
-	if instance.ENIId != "" && d.networkPlumber != nil {
-		// VPC mode: create tap device and add to OVS br-int
-		if err := d.networkPlumber.SetupTapDevice(instance.ENIId, instance.ENIMac); err != nil {
-			slog.Error("Failed to set up tap device", "eni", instance.ENIId, "err", err)
-			return fmt.Errorf("setup tap device: %w", err)
-		}
-
-		tapName := TapDeviceName(instance.ENIId)
-		instance.Config.NetDevs = append(instance.Config.NetDevs, vm.NetDev{
-			Value: fmt.Sprintf("tap,id=net0,ifname=%s,script=no,downscript=no", tapName),
-		})
-		instance.Config.Devices = append(instance.Config.Devices, vm.Device{
-			Value: fmt.Sprintf("virtio-net-pci,netdev=net0,mac=%s", instance.ENIMac),
-		})
-
-		slog.Info("VPC networking configured", "tap", tapName, "eni", instance.ENIId, "mac", instance.ENIMac)
-
-		// Additional VPC NICs for multi-subnet system VMs (e.g. ALBs with
-		// subnets across multiple AZs).
-		if err := d.setupExtraENINICs(instance); err != nil {
-			return err
-		}
-
-		// DEV_NETWORKING: add a second NIC with hostfwd for SSH dev access
-		if d.config.Daemon.DevNetworking {
-			sshDebugAddr, err := viperblock.FindFreePort()
-			if err != nil {
-				slog.Warn("DEV_NETWORKING: failed to find free port for dev NIC", "err", err)
-			} else {
-				_, sshDebugPort, err := net.SplitHostPort(sshDebugAddr)
-				if err != nil {
-					slog.Warn("DEV_NETWORKING: failed to parse port from address", "addr", sshDebugAddr, "err", err)
-				} else {
-					bindIP := d.config.Host
-					if bindIP == "" || bindIP == "0.0.0.0" {
-						bindIP = "127.0.0.1"
-					}
-					netdevVal := fmt.Sprintf("user,id=dev0,hostfwd=tcp:%s:%s-:22", bindIP, sshDebugPort)
-					// Add extra hostfwd rules for system instances (e.g. ALB VMs forwarding HTTP ports)
-					if instance.ExtraHostfwd != nil {
-						for guestPort := range instance.ExtraHostfwd {
-							fwdAddr, fwdErr := viperblock.FindFreePort()
-							if fwdErr != nil {
-								slog.Warn("DEV_NETWORKING: failed to find free port for extra hostfwd", "guestPort", guestPort, "err", fwdErr)
-								continue
-							}
-							_, hostPort, splitErr := net.SplitHostPort(fwdAddr)
-							if splitErr != nil {
-								slog.Warn("DEV_NETWORKING: failed to parse extra hostfwd address", "fwdAddr", fwdAddr, "err", splitErr)
-								continue
-							}
-							hostPortInt, convErr := strconv.Atoi(hostPort)
-							if convErr != nil {
-								slog.Warn("DEV_NETWORKING: failed to convert extra hostfwd port", "hostPort", hostPort, "err", convErr)
-								continue
-							}
-							netdevVal += fmt.Sprintf(",hostfwd=tcp:%s:%s-:%d", bindIP, hostPort, guestPort)
-							instance.ExtraHostfwd[guestPort] = hostPortInt
-							slog.Info("DEV_NETWORKING: extra hostfwd", "guestPort", guestPort, "hostPort", hostPort, "instanceId", instance.ID)
-						}
-					}
-					instance.Config.NetDevs = append(instance.Config.NetDevs, vm.NetDev{
-						Value: netdevVal,
-					})
-					devMac := generateDevMAC(instance.ID)
-					instance.Config.Devices = append(instance.Config.Devices, vm.Device{
-						Value: fmt.Sprintf("virtio-net-pci,netdev=dev0,mac=%s", devMac),
-					})
-					slog.Info("DEV_NETWORKING: added dev NIC with SSH hostfwd",
-						"bindIP", bindIP, "port", sshDebugPort, "mac", devMac, "instanceId", instance.ID)
-				}
-			}
-		}
-	} else {
-		// Non-VPC fallback: user-mode networking with SSH port forwarding
-		sshDebugAddr, err := viperblock.FindFreePort()
-		if err != nil {
-			slog.Error("Failed to find free port", "err", err)
-			return err
-		}
-		_, sshDebugPort, err := net.SplitHostPort(sshDebugAddr)
-		if err != nil {
-			slog.Error("Failed to parse port from address", "addr", sshDebugAddr, "err", err)
-			return fmt.Errorf("parse port from %s: %w", sshDebugAddr, err)
-		}
-
-		bindIP := d.config.Host
-		if bindIP == "" || bindIP == "0.0.0.0" {
-			bindIP = "127.0.0.1"
-		}
-		instance.Config.NetDevs = append(instance.Config.NetDevs, vm.NetDev{
-			Value: fmt.Sprintf("user,id=net0,hostfwd=tcp:%s:%s-:22", bindIP, sshDebugPort),
-		})
-		instance.Config.Devices = append(instance.Config.Devices, vm.Device{
-			Value: "virtio-net-pci,netdev=net0",
-		})
-	}
-
-	// Management NIC: system instances get a TAP on br-mgmt for control plane traffic.
-	if instance.MgmtMAC != "" && instance.MgmtTap != "" {
-		instance.Config.NetDevs = append(instance.Config.NetDevs, vm.NetDev{
-			Value: fmt.Sprintf("tap,id=mgmt0,ifname=%s,script=no,downscript=no", instance.MgmtTap),
-		})
-		instance.Config.Devices = append(instance.Config.Devices, vm.Device{
-			Value: fmt.Sprintf("virtio-net-pci,netdev=mgmt0,mac=%s", instance.MgmtMAC),
-		})
-		slog.Info("Management NIC configured", "tap", instance.MgmtTap, "mac", instance.MgmtMAC, "ip", instance.MgmtIP, "instanceId", instance.ID)
-	}
-
-	instance.Config.Devices = append(instance.Config.Devices, vm.Device{
-		Value: "virtio-rng-pci",
-	})
-
-	// QMP socket
-	qmpSocket, err := utils.GenerateSocketFile(fmt.Sprintf("qmp-%s", instance.ID))
-
-	if err != nil {
-		slog.Error("Failed to generate QMP socket", "err", err)
-		return err
-	}
-
-	instance.Config.QMPSocket = qmpSocket
-
-	// Temp, wait for nbdkit to start
-	// TODO: Improve, confirm nbdkit started for each volume
-	time.Sleep(2 * time.Second)
-
-	// Create a unique error channel for this specific mount request
-	processChan := make(chan int, 1)
-	exitChan := make(chan int, 1)
-	startupConfirmed := make(chan bool, 1)
-
-	go func() {
-		cmd, err := instance.Config.Execute()
-
-		if err != nil {
-			slog.Error("Failed to execute VM", "err", err)
-			processChan <- 0
-			return
-		}
-
-		VMstdout, err := cmd.StdoutPipe()
-		if err != nil {
-			slog.Error("Failed to pipe STDERR VM", "err", err)
-			processChan <- 0
-			return
-		}
-
-		VMstderr, err := cmd.StderrPipe()
-		if err != nil {
-			slog.Error("Failed to pipe STDERR VM", "err", err)
-			processChan <- 0
-			return
-		}
-
-		err = cmd.Start()
-
-		if err != nil {
-			slog.Error("Failed to start VM", "err", err)
-			processChan <- 0
-			return
-		}
-
-		slog.Info("VM started successfully", "pid", cmd.Process.Pid)
-
-		// Set OOM score for QEMU process (prefer killing VMs over system services)
-		if err := utils.SetOOMScore(cmd.Process.Pid, 500); err != nil {
-			slog.Warn("Failed to set QEMU OOM score", "pid", cmd.Process.Pid, "err", err)
-		}
-
-		// Log QEMU stdout (serial output is captured via chardev logfile, not stdout)
-		go func() {
-			scanner := bufio.NewScanner(VMstdout)
-			for scanner.Scan() {
-				slog.Info("[qemu]", "line", scanner.Text())
-			}
-		}()
-
-		// --- reader for STDERR ---
-		go func() {
-			scanner := bufio.NewScanner(VMstderr)
-			slog.Info("QEMU stderr reader started")
-
-			for scanner.Scan() {
-				line := scanner.Text()
-				slog.Error("[qemu-stderr]", "line", line)
-			}
-		}()
-
-		processChan <- cmd.Process.Pid
-
-		// Block until QEMU exits
-		waitErr := cmd.Wait()
-
-		if waitErr != nil {
-			slog.Error("VM process exited", "instance", instance.ID, "err", waitErr)
-		}
-
-		// Signal startup check (non-blocking)
-		select {
-		case exitChan <- 1:
-		default:
-		}
-
-		// Wait for startup phase to complete before deciding on crash handling
-		confirmed := <-startupConfirmed
-		if !confirmed {
-			return // Startup failed, LaunchInstance handles the error
-		}
-
-		// Handle exit: crash vs clean shutdown
-		if waitErr != nil {
-			d.handleInstanceCrash(instance, waitErr)
-		} else {
-			slog.Info("VM process exited cleanly", "instance", instance.ID)
-		}
-	}()
-
-	// Wait for startup result
-	pid := <-processChan
-
-	if pid == 0 {
-		return fmt.Errorf("failed to start qemu")
-	}
-
-	// Wait for 1 second to confirm nbdkit is running
-	time.Sleep(1 * time.Second)
-
-	// Check if QEMU exited immediately with an error
-	select {
-	case exitErr := <-exitChan:
-		startupConfirmed <- false // tell goroutine not to handle crash
-		if exitErr != 0 {
-			errorMsg := fmt.Errorf("failed: %v", exitErr)
-			slog.Error("Failed to launch qemu", "err", errorMsg)
-			return errorMsg
-		}
-	default:
-		startupConfirmed <- true // goroutine will handle future crashes
-		slog.Info("QEMU started successfully and is running",
-			"console_log", instance.Config.ConsoleLogPath,
-			"serial_socket", instance.Config.SerialSocket)
-	}
-
-	// Confirm the instance has booted
-	_, err = utils.ReadPidFile(instance.ID)
-
-	if err != nil {
-		slog.Error("Failed to read PID file", "err", err)
-		return err
-	}
-
-	return nil
-}
-
-// buildBaseVMConfig creates a vm.Config with base QEMU settings and PCIe
-// hotplug root ports. Architecture, vCPU, and memory come from the caller
-// (resolved from instance type info).
-func buildBaseVMConfig(instanceID, pidFile, consoleLogPath, serialSocket, architecture string, vCPUs, memoryMiB int) vm.Config {
-	cfg := vm.Config{
-		Name:           instanceID,
-		PIDFile:        pidFile,
-		EnableKVM:      true,
-		NoGraphic:      true,
-		MachineType:    "q35",
-		ConsoleLogPath: consoleLogPath,
-		SerialSocket:   serialSocket,
-		CPUType:        "host",
-		Memory:         memoryMiB,
-		CPUCount:       vCPUs,
-		Architecture:   architecture,
-	}
-
-	// Add PCIe root ports for volume hotplug (Q35 requires explicit root ports).
-	// 11 ports for /dev/sd[f-p] hotplug slots, starting at chassis 1.
-	for i := 1; i <= 11; i++ {
-		cfg.Devices = append(cfg.Devices, vm.Device{
-			Value: fmt.Sprintf("pcie-root-port,id=hotplug%d,chassis=%d,slot=0", i, i),
-		})
-	}
-
-	return cfg
-}
-
-// buildDrives converts EBS volume requests into QEMU drive, iothread, and device
-// configurations. Returns an error if any non-EFI volume is missing its NBDURI.
-func buildDrives(requests []types.EBSRequest, cpuCount int) ([]vm.Drive, []vm.IOThread, []vm.Device, error) {
-	var drives []vm.Drive
-	var iothreads []vm.IOThread
-	var devices []vm.Device
-
-	for _, v := range requests {
-		// TODO: Add EFI support
-		if v.EFI {
-			continue
-		}
-
-		if v.NBDURI == "" {
-			return nil, nil, nil, fmt.Errorf("NBDURI not set for volume %s - was volume mounted?", v.Name)
-		}
-
-		drive := vm.Drive{File: v.NBDURI}
-
-		if v.Boot {
-			drive.Format = "raw"
-			drive.If = "none"
-			drive.Media = "disk"
-			drive.ID = "os"
-			drive.Cache = "none"
-
-			iothreadID := "ioth-os"
-			iothreads = append(iothreads, vm.IOThread{ID: iothreadID})
-			devices = append(devices, vm.Device{
-				Value: fmt.Sprintf("virtio-blk-pci,drive=%s,iothread=%s,num-queues=%d,bootindex=1",
-					drive.ID, iothreadID, cpuCount),
-			})
-		}
-
-		if v.CloudInit {
-			drive.Format = "raw"
-			drive.If = "virtio"
-			drive.Media = "cdrom"
-			drive.ID = "cloudinit"
-		}
-
-		slog.Info("Using NBD URI for drive", "volume", v.Name, "uri", v.NBDURI)
-		drives = append(drives, drive)
-	}
-
-	return drives, iothreads, devices, nil
 }
 
 // ebsTopic returns a node-specific EBS NATS topic, e.g. "ebs.node1.mount".
